restructure
Поменял структуру проекта чтобы соответствовать стандартам
