+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.activities.NutritionActivity"
+            android:name=".ui.domain.NutritionActivity"
             android:exported="false" />
     </application>
