+++ app/src/main/java/com/example/happ_frontend/MainActivity.kt
@@ -15,7 +15,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.happ_frontend.ui.activities.NutritionActivity
+import com.example.happ_frontend.ui.domain.NutritionActivity
 import com.example.happ_frontend.ui.theme.HappfrontendTheme
 
 class MainActivity : ComponentActivity() {
