+++ app/src/main/java/com/example/happ_frontend/ui/screens/nutrition/NutritionScreen.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.happ_frontend.ui.viewmodels.NutritionViewModel
+import com.example.happ_frontend.model.nutrition.NutritionViewModel
 
 @Composable
 fun NutritionScreen(
